fixed errors and added new functions

+++ ft_strlen.c
@@ -6,11 +6,13 @@
 /*   By: sschelti <sschelti@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/06 17:06:09 by sschelti          #+#    #+#             */
-/*   Updated: 2022/10/11 17:38:28 by sschelti         ###   ########.fr       */
+/*   Updated: 2022/10/14 17:48:45 by sschelti         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_strlen(const char *str)
+#include "libft.h"
+
+size_t	ft_strlen(const char *str)
 {
 	int	i;
